requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langchain-mcp-adapters>=0.1.0
fastmcp>=2.0.0
chromadb>=0.5.0
streamlit>=1.38.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
langchain-anthropic>=0.1.0
